An experiment of mine: The ArraySet
This is an implementation of the SortedSet interface. Instances of ArraySet class use an internal array to store the elements. There is no length limit because they can grow dinamically and their elements are sorted by a comparator that can be changed, even after have been created. The "grow policy" can be set any time by its capacity.
Besides, as an ArraySet uses arrays, it allows direct access to its elements by index.